+++ openc3/lib/openc3/api/tlm_api.rb
@@ -25,7 +25,7 @@ require 'openc3/models/cvt_model'
 require 'openc3/packets/packet'
 require 'openc3/topics/telemetry_topic'
 require 'openc3/topics/interface_topic'
-require 'openc3/topics/decom_topic'
+require 'openc3/topics/telemetry_decom_topic'
 
 module OpenC3
   module Api
@@ -150,7 +150,7 @@ module OpenC3
       if interface_name
         InterfaceTopic.inject_tlm(interface_name, target_name, packet_name, item_hash, type: type, scope: scope)
       else
-        DecomTopic.inject_tlm(target_name, packet_name, item_hash, type: type, scope: scope)
+        TelemetryDecomTopic.inject_tlm(target_name, packet_name, item_hash, type: type, scope: scope)
       end
     end
